Solo falta agregar que se genere la contraseña aleatoria

+++ routes/password.py
@@ -18,7 +18,7 @@ def get_passwords(current_user: User = Depends(get_current_active_user), db: Ses
         raise HTTPException(status_code=401, detail="Unauthenticated user")
 
     # Filtrar las contraseñas del usuario activo
-    user_passwords = db.query(Password).filter(Password.user_id == current_user.id).all()
+    user_passwords = db.query(PasswordModel).filter(PasswordModel.user_id == current_user.id).all()
     return user_passwords
 
 @password_router.get("/passwords/{id}", tags=['passwords'], response_model = Password, status_code=200)
@@ -37,16 +37,18 @@ def create_password(password: Password, current_user: User = Depends(get_current
         raise HTTPException(status_code=401, detail="Unauthenticated user")
 
     # Crea una nueva contraseña asociada al usuario actual
-    new_password = Password(**password.model_dump(), user_id=current_user.id)
+    new_password = PasswordModel(**password.model_dump(), user_id=current_user.id)
     db.add(new_password)
     db.commit()
     return JSONResponse(status_code=201, content={"message": "Password created successfully"})
 
 @password_router.put("/passwords/{id}", tags=['passwords'], response_model = dict, status_code=200)
-def update_password(id: int, password: Password):
+def update_password(id: int, password: Password, current_user: User = Depends(get_current_active_user)):
     db = Session()
     result = db.query(PasswordModel).filter(PasswordModel.id == id).first()
     if result:
+        if result.user_id != current_user.id:
+            raise HTTPException(status_code=403, detail="Not authorized to update this password")
         result.userName = password.userName
         result.userEmail = password.userEmail
         result.password = password.password
@@ -58,10 +60,12 @@ def update_password(id: int, password: Password):
     return response
 
 @password_router.delete("/passwords/{id}", tags=['passwords'], response_model = dict, status_code=200)
-def delete_password(id: int):
+def delete_password(id: int, current_user: User = Depends(get_current_active_user)):
     db = Session()
     result = db.query(PasswordModel).filter(PasswordModel.id == id).first()
     if result:
+        if result.user_id != current_user.id:
+            raise HTTPException(status_code=403, detail="Not authorized to delete this password")
         db.delete(result)
         db.commit()
         response = JSONResponse(status_code=200, content={"message": "Password deleted successfully"})
